Fixed problems with the speed script since im a dumbass and had to rewrite the entire thing a few times before i figured it out...

+++ Assets/Motion Sickness/Scripts/SpeedVsDistance.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public class SpeedVsDistance : MonoBehaviour
 {
-    public int[] distance;
-    public int[] speed;
+    public float[] distance;
+    public float[] speed;
 }
